working makefile with multiple includes
