+++ etl/population.mk
@@ -1,0 +1,2 @@
+output/populations.json: 
+	python3 processors/get_county_populations.py > $@
